Move freestanding entities out of the documents tests so that the mock works.
